refactor: move optimized-resize to core, add `debouncedResize` event
