+++ packages/components/bolt-tabs/src/tabs.js
@@ -14,6 +14,8 @@ import classNames from 'classnames/bind';
 import styles from './tabs.scss';
 import schema from '../tabs.schema.yml';
 
+import '@bolt/core/utils/optimized-resize';
+
 // define which specific props to provide to children that subscribe
 export const TabsContext = defineContext({
   inset: 'auto',
@@ -537,7 +539,7 @@ class BoltTabs extends withContext(withLitHtml()) {
         this._resizeMenu();
       });
 
-      window.addEventListener('optimizedResize', this._resizeMenu);
+      window.addEventListener('throttledResize', this._resizeMenu);
       this.dropdownButton.addEventListener('click', this._handleDropdownToggle);
 
       this.ready = true;
@@ -556,7 +558,7 @@ class BoltTabs extends withContext(withLitHtml()) {
     this.ready = false;
     this.removeAttribute('ready');
 
-    window.removeEventListener('optimizedResize', this._resizeMenu);
+    window.removeEventListener('throttledResize', this._resizeMenu);
     document.removeEventListener('click', this._handleExternalClicks);
     this.dropdownButton &&
       this.dropdownButton.removeEventListener(
@@ -586,31 +588,5 @@ class BoltTabs extends withContext(withLitHtml()) {
     `;
   }
 }
-
-// Create a custom 'optimizedResize' event that works just like window.resize but is more performant because it
-// won't fire before a previous event is complete.
-// This was adapted from https://developer.mozilla.org/en-US/docs/Web/Events/resize
-(function() {
-  function throttle(type, name, obj) {
-    obj = obj || window;
-    let running = false;
-
-    function func() {
-      if (running) {
-        return;
-      }
-      running = true;
-      requestAnimationFrame(function() {
-        obj.dispatchEvent(new CustomEvent(name));
-        running = false;
-      });
-    }
-    obj.addEventListener(type, func);
-  }
-
-  // Initialize on window.resize event.  Note that throttle can also be initialized on any type of event,
-  // such as scroll.
-  throttle('resize', 'optimizedResize');
-})();
 
 export { BoltTabs };
